Remove boolean workaround
Chaiscript's [bug](https://github.com/ChaiScript/ChaiScript/issues/481) with functions returning boolean which are producing
an exception seems to be resolved (with [PR #503](https://github.com/ChaiScript/ChaiScript/pull/503)).

For now I've applied this patch in a fork of ChaiScript repo and replaced
dependency from this repository in order to pointing to that fixed version.

Once ChaiScript will be free of this bug, I'll change dependency for pointing to
the original repository.

Removed workaround and adjusted all unittests and other code.